Add test script
This motivation for this was because Copilot was
creating commits which would fail to deploy to
Convex which made me realise there was no way to
get quick feedback on basic type errors.

Hopefully it will think to use the test script
without explicit prompting seen as its a common
convention and use the feedback to iterate better.